+++ services/__init__.py
@@ -1,3 +1,3 @@
-from .file_service import *
+from .file_service import FileService
 from .tyre_model_service import TyreModelService
 from .tyre_impression_service import TyreImpressionService